+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * Added `ktorJvmEngine` parameter to plugin with option to choose ktor engine.
 * Added `onHandlerException` for spring starter bot configuration, to cover exceptions thrown by update handler.
-* Added extenstion for `String` for escaping formatting, `escapeFormatting(parseMode)`, `escapeMarkdown`,
+* Added extension for `String` for escaping formatting, `escapeFormatting(parseMode)`, `escapeMarkdown`,
   `escapeMarkdownV2`, `escapeHTML`.
 
 ### 7.6.2